backend/requirements.txt
fastapi>=0.108.0,<1.0.0
uvicorn[standard]>=0.25.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
alembic>=1.13.0,<2.0.0
pydantic>=2.5.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0
redis>=5.0.0,<6.0.0
anthropic>=0.18.0,<1.0.0
google-generativeai>=0.3.0,<1.0.0
httpx>=0.26.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
google-auth>=2.25.2
google-cloud-aiplatform>=1.38.1
